wrapper/requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv
boto3